# binno_backend/contracts/binno-openapi-v1.yaml
openapi: 3.1.0
info:
  title: BINNO Platform API
  version: 1.1.0
  description: |
    BINNO MVP contract-first API. All IDs are UUID v7, all timestamps are
    UTC, all money is integer tiyin. Mutations require an Idempotency-Key.
servers:
  - url: https://api.binno.uz/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /orders/{id}/feedback:
    post:
      tags: [trust]
      summary: Binary feedback on a closed order
      operationId: createOrderFeedback
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderFeedbackInput' }
      responses:
        '201': { description: Feedback recorded }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. Required claims:
          * `sub`: the user's UUID
          * `exp`: mandatory; a token without exp is rejected outright
          * `roles`: array of `buyer` | `seller` | `operator`;
            unknown roles are dropped

        Authentication (is the token genuine) and authorisation (is this
        resource yours) are separate concerns. A genuine token is not enough:
        every endpoint checks the resource's owner on its own. Requesting
        another user's order or owner account answers 403, or 404 where
        existence must not be disclosed.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
      description: |
        Retry-safe mutation key with a 24-hour window.

        Same key + same body -> the first response is replayed verbatim
        (`Idempotency-Replayed: true`).
        Same key + a DIFFERENT body -> 422, `type` =
        `https://binno.uz/problems/idempotency-key-reuse`. The mutation does
        not run. Use a fresh key for every logically distinct request.
...
  responses:
    Problem:
      description: RFC 7807 error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Illegal state transition or duplicate action
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Validation:
      description: Validation error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Token missing, malformed, or expired
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: |
        The token is genuine but the resource belongs to another account.
        Where disclosing existence is a risk (for example someone else's
        order), 404 is returned instead.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: |
        The caller's request budget is exhausted. Authenticated callers are
        counted per principal, anonymous callers per IP, and the anonymous
        budget is smaller.
      headers:
        Retry-After:
          description: Seconds until the next window
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    OrderFeedbackInput:
      type: object
      required: [had_problem, channel]
      properties:
        had_problem: { type: boolean }
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [late, qty_short, quality, price_mismatch, no_response]
        comment: { type: string, maxLength: 2000 }
        channel: { type: string, enum: [push, sms, app] }
...